refactor: extract shared UpdateLogTable component from Dashboard and History
Closes #31

Both pages rendered the same expandable update log table (~130 lines each):
ImageVersionLabel, expandedRows state, parsedDiff JSON parsing, DiffBadge
dialog, and the full Table/Fragment structure.

Extract into ui/src/components/history/UpdateLogTable.tsx. Both pages are
now single-line consumers:

  <UpdateLogTable entries={history?.data ?? []} emptyMessage="..." />

History.tsx retains its filters, stats, and pagination. Dashboard.tsx
retains its agent grid and activity heading.
